+++ src/main/java/com/health/healthlog/repository/TrainingRepository.java
@@ -3,5 +3,5 @@ package com.health.healthlog.repository;
 import com.health.healthlog.domain.Training;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TrainingRepository extends JpaRepository<Training, Integer> {
+public interface TrainingRepository extends JpaRepository<Training, Long> {
 }
